making flask api more restful
