+++ Webapp/sources/blueprints/api/sections/__init__.py
@@ -1,0 +1,5 @@
+from apiflask import APIBlueprint
+
+sections_api_bp = APIBlueprint("sections", __name__, url_prefix="/sections")
+
+from . import routes
