+++ Webapp/sources/blueprints/api/templates/__init__.py
@@ -1,0 +1,5 @@
+from apiflask import APIBlueprint
+
+templates_api_bp = APIBlueprint("templates", __name__, url_prefix="/templates")
+
+from . import routes
